requirements.txt
aws-cdk-lib==2.229.1
constructs>=10.0.0,<11.0.0
cdk-nag>=2.0.0
aws-cdk.aws-lambda-python-alpha>=2.229.1a0
